Fix type export
