+++ frontend/src/metabase/components/form/CustomForm/CustomFormFooter/CustomFormFooterTypes.ts
@@ -1,0 +1,8 @@
+export interface CustomFormFooterProps {
+  submitTitle: string;
+  cancelTitle?: string;
+  fullWidth?: boolean;
+  isModal?: boolean;
+  footerExtraButtons: React.ReactElement[];
+  onCancel?: () => void;
+}
