+++ frontend/src/metabase/components/form/CustomForm/CustomFormFooter/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./CustomFormFooter";
+export * from "./CustomFormFooterTypes";
